+++ src/main/java/io/supertokens/webserver/api/thirdparty/GetUsersByEmailAPI.java
@@ -21,7 +21,6 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.supertokens.Main;
-import io.supertokens.exceptions.InvalidInputException;
 import io.supertokens.pluginInterface.RECIPE_ID;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.thirdparty.UserInfo;
@@ -35,6 +34,8 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class GetUsersByEmailAPI extends WebserverAPI {
+    private static final long serialVersionUID = -4413719941975228004L;
+
     public GetUsersByEmailAPI(Main main) {
         super(main, RECIPE_ID.THIRD_PARTY.toString());
     }
@@ -57,9 +58,6 @@ public class GetUsersByEmailAPI extends WebserverAPI {
             result.add("users", usersJson);
 
             super.sendJsonResponse(200, result, resp);
-        } catch (InvalidInputException e) {
-            throw new ServletException(
-                    new BadRequestException(e.getMessage()));
         } catch (StorageQueryException e) {
             throw new ServletException(e);
         }