+++ CHANGELOG.md
@@ -15,6 +15,11 @@ https://github.com/supertokens/supertokens-core/issues/255
 - Add `GET /recipe/users/by-email?email=john@example.com` endpoint for ThirdParty recipe to fetch all users with given
  email
 
+## [3.4.2] - 2021-06-27
+### Fixes
+- `NullPointerException` that is thrown in `AccessTokenSigningKey.java` class when the `keyInfo` object is accessed
+ in parallel after the signing key has expired: https://github.com/supertokens/supertokens-core/issues/282
+
 ## [3.4.1] - 2021-06-18
 ### Added
 - `test_mode` to the options for running the core so that it can be run in test mode whilst being tested by the